+++ evm/rust/Cargo.toml
@@ -1,15 +1,14 @@
 [package]
-name = "ismp-solidity-abi"
+name = "ismp-abi"
 version = "0.1.1"
 edition = "2021"
 authors = ["Polytope Labs <hello@polytope.technology>"]
 description = "Generated rust types for the ISMP solidity ABI"
-publish = false
+repository = "https://github.com/polytope-labs/hyperbridge"
 
 [dependencies]
 primitive-types = { workspace = true, default-features = false }
 anyhow = { workspace = true, default-features = false }
-ismp = { workspace = true, default-features = false }
 
 alloy-sol-types = { workspace = true, default-features = false }
 alloy-sol-macro = { workspace = true }
@@ -37,7 +36,6 @@ std = [
     "dep:alloy-transport",
     "anyhow/std",
     "primitive-types/std",
-    "ismp/std",
     "alloy-sol-types/std",
     "alloy-primitives/std",
     "beefy-verifier-primitives/std",
