+++ evm/rust/src/conversions.rs
@@ -15,32 +15,13 @@
 
 //! Convenient type conversions
 
-use crate::{
-	ecdsa_beefy::Beefy::IntermediateState,
-	evm_host::EvmHost::{
-		GetRequest, GetRequestEvent, GetRequestHandled, GetRequestTimeoutHandled, GetResponse,
-		HostFrozen, HostParamsUpdated, HostWithdrawal, PostRequest, PostRequestEvent,
-		PostRequestHandled, PostRequestTimeoutHandled, PostResponse, PostResponseEvent,
-		PostResponseFunded, PostResponseHandled, PostResponseTimeoutHandled, RequestFunded,
-		StateCommitment, StateCommitmentRead, StateCommitmentVetoed as EvmStateCommitmentVetoed,
-		StateMachineHeight, StateMachineUpdated as EvmStateMachineUpdated,
-	},
-};
+use crate::ecdsa_beefy::Beefy::IntermediateState;
 
-use alloc::string::{String, ToString};
-use alloy_primitives::{FixedBytes, U256};
-use anyhow::anyhow;
-use core::str::FromStr;
-use ismp::{
-	consensus::StateMachineId,
-	events::{StateCommitmentVetoed, StateMachineUpdated, TimeoutHandled},
-	host::StateMachine,
-	router,
-};
-use primitive_types::{H160, H256};
+use alloy_primitives::U256;
+use primitive_types::H256;
 
 /// Helper trait for converting primitive types to alloy U256
-trait ToU256 {
+pub trait ToU256 {
 	fn to_u256(self) -> U256;
 }
 
@@ -449,315 +430,6 @@ impl From<IntermediateState> for local::IntermediateState {
 				overlay_root: H256(value.commitment.overlayRoot.0),
 			},
 		}
-	}
-}
-
-impl From<router::PostResponse> for PostResponse {
-	fn from(value: router::PostResponse) -> Self {
-		PostResponse {
-			request: value.post.into(),
-			response: value.response.into(),
-			timeoutTimestamp: value.timeout_timestamp,
-		}
-	}
-}
-
-impl From<router::PostRequest> for PostRequest {
-	fn from(value: router::PostRequest) -> Self {
-		PostRequest {
-			source: value.source.to_string().into_bytes().into(),
-			dest: value.dest.to_string().into_bytes().into(),
-			nonce: value.nonce,
-			from: value.from.into(),
-			to: value.to.into(),
-			timeoutTimestamp: value.timeout_timestamp,
-			body: value.body.into(),
-		}
-	}
-}
-
-impl From<router::PostRequest> for crate::handler::Handler::PostRequest {
-	fn from(value: router::PostRequest) -> Self {
-		Self {
-			source: value.source.to_string().into_bytes().into(),
-			dest: value.dest.to_string().into_bytes().into(),
-			nonce: value.nonce,
-			from: value.from.into(),
-			to: value.to.into(),
-			timeoutTimestamp: value.timeout_timestamp,
-			body: value.body.into(),
-		}
-	}
-}
-
-impl From<router::PostResponse> for crate::handler::Handler::PostResponse {
-	fn from(value: router::PostResponse) -> Self {
-		Self {
-			request: value.post.into(),
-			response: value.response.into(),
-			timeoutTimestamp: value.timeout_timestamp,
-		}
-	}
-}
-
-// HandlerV2 carries its own distinct `sol!`-generated `PostRequest` /
-// `PostResponse` types (same ABI shape as V1, different Rust type identity
-// since they live in a separate `sol!` module). Provide the same conversions
-// so callers submitting through `HandlerV2Instance` can reuse the ISMP
-// `router::Post{Request,Response}` as-is.
-impl From<router::PostRequest> for crate::handler::handler_v2::HandlerV2::PostRequest {
-	fn from(value: router::PostRequest) -> Self {
-		Self {
-			source: value.source.to_string().into_bytes().into(),
-			dest: value.dest.to_string().into_bytes().into(),
-			nonce: value.nonce,
-			from: value.from.into(),
-			to: value.to.into(),
-			timeoutTimestamp: value.timeout_timestamp,
-			body: value.body.into(),
-		}
-	}
-}
-
-impl From<router::PostResponse> for crate::handler::handler_v2::HandlerV2::PostResponse {
-	fn from(value: router::PostResponse) -> Self {
-		Self {
-			request: value.post.into(),
-			response: value.response.into(),
-			timeoutTimestamp: value.timeout_timestamp,
-		}
-	}
-}
-
-impl TryFrom<PostRequest> for router::PostRequest {
-	type Error = anyhow::Error;
-	fn try_from(value: PostRequest) -> Result<Self, Self::Error> {
-		Ok(router::PostRequest {
-			source: StateMachine::from_str(
-				&String::from_utf8(value.source.to_vec()).map_err(|e| anyhow!("{e}"))?,
-			)
-			.map_err(|err| anyhow!("{err}"))?,
-			dest: StateMachine::from_str(
-				&String::from_utf8(value.dest.to_vec()).map_err(|e| anyhow!("{e}"))?,
-			)
-			.map_err(|err| anyhow!("{err}"))?,
-			nonce: value.nonce.try_into().map_err(|e| anyhow!("{e}"))?,
-			from: value.from.to_vec(),
-			to: value.to.to_vec(),
-			timeout_timestamp: value.timeoutTimestamp.try_into().map_err(|e| anyhow!("{e}"))?,
-			body: value.body.to_vec(),
-		})
-	}
-}
-
-impl From<router::GetRequest> for GetRequest {
-	fn from(value: router::GetRequest) -> Self {
-		GetRequest {
-			source: value.source.to_string().into_bytes().into(),
-			dest: value.dest.to_string().into_bytes().into(),
-			nonce: value.nonce,
-			keys: value.keys.into_iter().map(Into::into).collect(),
-			from: {
-				let mut address = H160::default();
-				address.0.copy_from_slice(&value.from);
-				alloy_primitives::Address::from(address.0)
-			},
-			context: value.context.into(),
-			timeoutTimestamp: value.timeout_timestamp,
-			height: value.height,
-		}
-	}
-}
-
-impl From<router::GetResponse> for GetResponse {
-	fn from(value: router::GetResponse) -> Self {
-		GetResponse {
-			request: value.get.into(),
-			values: value
-				.values
-				.into_iter()
-				.map(|storage_value| crate::evm_host::StorageValue {
-					key: storage_value.key.into(),
-					value: storage_value.value.unwrap_or_default().into(),
-				})
-				.collect(),
-		}
-	}
-}
-
-impl TryFrom<ismp::consensus::StateMachineHeight> for StateMachineHeight {
-	type Error = anyhow::Error;
-	fn try_from(value: ismp::consensus::StateMachineHeight) -> Result<Self, anyhow::Error> {
-		Ok(StateMachineHeight {
-			stateMachineId: match value.id.state_id {
-				StateMachine::Polkadot(id) | StateMachine::Kusama(id) => id.to_u256(),
-				state_machine => Err(anyhow!("Unsupported state machine {state_machine:?}"))?,
-			},
-			height: value.height.to_u256(),
-		})
-	}
-}
-
-impl From<ismp::consensus::StateCommitment> for StateCommitment {
-	fn from(value: ismp::consensus::StateCommitment) -> Self {
-		StateCommitment {
-			timestamp: value.timestamp.to_u256(),
-			stateRoot: FixedBytes::from(value.state_root.0),
-			overlayRoot: FixedBytes::from(value.overlay_root.unwrap_or_default().0),
-		}
-	}
-}
-
-/// Enum representing all EvmHost events for conversion
-pub enum EvmHostEvents {
-	GetRequestEvent(GetRequestEvent),
-	PostRequestEvent(PostRequestEvent),
-	PostResponseEvent(PostResponseEvent),
-	PostRequestHandled(PostRequestHandled),
-	GetRequestHandled(GetRequestHandled),
-	PostResponseHandled(PostResponseHandled),
-	StateMachineUpdated(EvmStateMachineUpdated),
-	PostRequestTimeoutHandled(PostRequestTimeoutHandled),
-	PostResponseTimeoutHandled(PostResponseTimeoutHandled),
-	GetRequestTimeoutHandled(GetRequestTimeoutHandled),
-	StateCommitmentVetoed(EvmStateCommitmentVetoed),
-	StateCommitmentRead(StateCommitmentRead),
-	HostFrozen(HostFrozen),
-	HostWithdrawal(HostWithdrawal),
-	HostParamsUpdated(HostParamsUpdated),
-	PostResponseFunded(PostResponseFunded),
-	RequestFunded(RequestFunded),
-}
-
-impl TryFrom<EvmHostEvents> for ismp::events::Event {
-	type Error = anyhow::Error;
-	fn try_from(event: EvmHostEvents) -> Result<Self, Self::Error> {
-		match event {
-			EvmHostEvents::GetRequestEvent(get) =>
-				Ok(ismp::events::Event::GetRequest(get.try_into()?)),
-			EvmHostEvents::PostRequestEvent(post) =>
-				Ok(ismp::events::Event::PostRequest(post.try_into()?)),
-			EvmHostEvents::PostResponseEvent(resp) =>
-				Ok(ismp::events::Event::PostResponse(router::PostResponse {
-					post: router::PostRequest {
-						source: StateMachine::from_str(&resp.dest).map_err(|e| anyhow!("{}", e))?,
-						dest: StateMachine::from_str(&resp.source).map_err(|e| anyhow!("{}", e))?,
-						nonce: resp.nonce.try_into().map_err(|e| anyhow!("{e}"))?,
-						from: resp.to.0.to_vec(),
-						to: resp.from.0.to_vec(),
-						timeout_timestamp: resp
-							.timeoutTimestamp
-							.try_into()
-							.map_err(|e| anyhow!("{e}"))?,
-						body: resp.body.to_vec(),
-					},
-					response: resp.response.to_vec(),
-					timeout_timestamp: resp
-						.responseTimeoutTimestamp
-						.try_into()
-						.map_err(|e| anyhow!("{e}"))?,
-				})),
-			EvmHostEvents::PostRequestHandled(handled) =>
-				Ok(ismp::events::Event::PostRequestHandled(ismp::events::RequestResponseHandled {
-					commitment: H256(handled.commitment.0),
-					relayer: handled.relayer.0.to_vec(),
-				})),
-			EvmHostEvents::GetRequestHandled(handled) =>
-				Ok(ismp::events::Event::GetRequestHandled(ismp::events::RequestResponseHandled {
-					commitment: H256(handled.commitment.0),
-					relayer: handled.relayer.0.to_vec(),
-				})),
-
-			EvmHostEvents::PostResponseHandled(handled) =>
-				Ok(ismp::events::Event::PostResponseHandled(ismp::events::RequestResponseHandled {
-					commitment: H256(handled.commitment.0),
-					relayer: handled.relayer.0.to_vec(),
-				})),
-			EvmHostEvents::StateMachineUpdated(filter) =>
-				Ok(ismp::events::Event::StateMachineUpdated(StateMachineUpdated {
-					state_machine_id: ismp::consensus::StateMachineId {
-						state_id: StateMachine::from_str(&filter.stateMachineId)
-							.map_err(|e| anyhow!("{}", e))?,
-						consensus_state_id: Default::default(),
-					},
-					latest_height: filter.height.try_into().map_err(|e| anyhow!("{e}"))?,
-				})),
-			EvmHostEvents::PostRequestTimeoutHandled(handled) => {
-				let dest = StateMachine::from_str(&handled.dest).map_err(|e| anyhow!("{}", e))?;
-				Ok(ismp::events::Event::PostRequestTimeoutHandled(TimeoutHandled {
-					commitment: H256(handled.commitment.0),
-					dest: dest.clone(),
-					source: dest.clone(),
-				}))
-			},
-			EvmHostEvents::PostResponseTimeoutHandled(handled) => {
-				let dest = StateMachine::from_str(&handled.dest).map_err(|e| anyhow!("{}", e))?;
-				Ok(ismp::events::Event::PostResponseTimeoutHandled(TimeoutHandled {
-					commitment: H256(handled.commitment.0),
-					dest: dest.clone(),
-					source: dest.clone(),
-				}))
-			},
-			EvmHostEvents::GetRequestTimeoutHandled(handled) => {
-				let dest = StateMachine::from_str(&handled.dest).map_err(|e| anyhow!("{}", e))?;
-				Ok(ismp::events::Event::GetRequestTimeoutHandled(TimeoutHandled {
-					commitment: H256(handled.commitment.0),
-					dest: dest.clone(),
-					source: dest.clone(),
-				}))
-			},
-			EvmHostEvents::StateCommitmentVetoed(vetoed) =>
-				Ok(ismp::events::Event::StateCommitmentVetoed(StateCommitmentVetoed {
-					height: ismp::consensus::StateMachineHeight {
-						id: StateMachineId {
-							state_id: StateMachine::from_str(&vetoed.stateMachineId)
-								.map_err(|e| anyhow!("{}", e))?,
-							consensus_state_id: Default::default(),
-						},
-						height: vetoed.height.try_into().map_err(|e| anyhow!("{e}"))?,
-					},
-					fisherman: vetoed.fisherman.0.to_vec(),
-				})),
-			EvmHostEvents::StateCommitmentRead(_) |
-			EvmHostEvents::HostFrozen(_) |
-			EvmHostEvents::HostWithdrawal(_) |
-			EvmHostEvents::HostParamsUpdated(_) |
-			EvmHostEvents::PostResponseFunded(_) |
-			EvmHostEvents::RequestFunded(_) => Err(anyhow!("Unsupported Event!"))?,
-		}
-	}
-}
-
-impl TryFrom<PostRequestEvent> for router::PostRequest {
-	type Error = anyhow::Error;
-
-	fn try_from(post: PostRequestEvent) -> Result<Self, Self::Error> {
-		Ok(router::PostRequest {
-			source: StateMachine::from_str(&post.source).map_err(|e| anyhow!("{}", e))?,
-			dest: StateMachine::from_str(&post.dest).map_err(|e| anyhow!("{}", e))?,
-			nonce: post.nonce.try_into().map_err(|e| anyhow!("{e}"))?,
-			from: post.from.0.to_vec(),
-			to: post.to.0.to_vec(),
-			timeout_timestamp: post.timeoutTimestamp.try_into().map_err(|e| anyhow!("{e}"))?,
-			body: post.body.to_vec(),
-		})
-	}
-}
-
-impl TryFrom<GetRequestEvent> for router::GetRequest {
-	type Error = anyhow::Error;
-
-	fn try_from(get: GetRequestEvent) -> Result<Self, Self::Error> {
-		Ok(router::GetRequest {
-			source: StateMachine::from_str(&get.source).map_err(|e| anyhow!("{}", e))?,
-			dest: StateMachine::from_str(&get.dest).map_err(|e| anyhow!("{}", e))?,
-			nonce: get.nonce.try_into().map_err(|e| anyhow!("{e}"))?,
-			from: get.from.0.to_vec(),
-			keys: get.keys.into_iter().map(|key| key.to_vec()).collect(),
-			height: get.height.try_into().map_err(|e| anyhow!("{e}"))?,
-			context: get.context.to_vec(),
-			timeout_timestamp: get.timeoutTimestamp.try_into().map_err(|e| anyhow!("{e}"))?,
-		})
 	}
 }
