+++ tesseract/consensus/integration-tests/Cargo.toml
@@ -21,7 +21,7 @@ alloy = { workspace = true, features = ["providers", "signer-local", "network", 
 tracing-subscriber = { version = "0.3.18", features = ["env-filter"] }
 
 merkle-mountain-range = { workspace = true }
-ismp-solidity-abi = { workspace = true }
+ismp-abi = { workspace = true }
 pallet-ismp-demo = { workspace = true }
 ismp = { workspace = true }
 sp-core = { workspace = true }
