+++ tesseract/messaging/messaging/Cargo.toml
@@ -28,7 +28,7 @@ tesseract-tron = { workspace = true }
 tesseract-evm = { workspace = true }
 subxt-utils = { workspace = true, default-features = true }
 pallet-ismp-relayer = { workspace = true, default-features = true }
-ismp-solidity-abi = { workspace = true, default-features = true }
+ismp-abi = { workspace = true, default-features = true }
 alloy-sol-types = { workspace = true, default-features = true }
 beefy-verifier-primitives = { workspace = true, default-features = true }
 codec = { workspace = true, default-features = true, features = ["derive"] }
