+++ tesseract/messaging/primitives/Cargo.toml
@@ -17,7 +17,7 @@ serde = { version = "1.0.164", features = ["derive"] }
 primitive-types =  { workspace = true, default-features = true }
 tracing = "0.1.40"
 parity-scale-codec = "3.2.2"
-ismp-solidity-abi = { workspace = true, default-features = true }
+ismp-abi = { workspace = true, default-features = true }
 
 pallet-ismp-relayer = { workspace = true, default-features = true }
 ismp = { workspace = true, default-features = true }
